Keyword-selection tweaks and a bugfix.
Played around a bit with YAKE options.
Found that 5 lines/keyword seemed reasonable for Wikipedia article content.
Skipping the first 2 lines of the content text also seemed to help, since [0:1] is usually the delimiter and the title, which is generally not something you want an outbound link made from.
Still, not really too impressed with the results.