Fix to keep frontmatter fields in order.
Previously, the existing document fields and the taxonomy fields were combined and deduplicated via list(set().union(set())), which was clever but did not preserve order.
Now, it starts with the taxonomy fields and adds the doc fields to the bottom, but only if they don't already exist in the taxonomy.

So in general, the taxonomy field ordering will dominate (given a taxonomy that covers most fields in the doc), which is probably desirable.