Add plaintext wordlist file for PyCharm spellchecker.
Doesn't appear to be any easy way to share the PyCharm project-level spelling dictionary in its default location (.idea/dictionaries/) and format (XML) without leaking the username of the local account running PyCharm. Not ideal.
So instead, this wordlist needs to be added manually to PyCharm as an additional wordlist (on each workstation), and words will have to be periodically added manually by copying over the contents of the project or app-level dictionary and stripping XML tags.