WIP: Continued working on obs-filter-vault.py (formerly obsidian-filter-vault.py).
Renamed to make consistent with other user tools, and also just to make the command less obnoxious to type if you have several of them on your PATH.
Also implemented a bunch of changes/improvements, but it still needs more work. Right now it handles trivial test cases, but not even simple real-world Obsidian vaults, due to the lack of exclusion-list capability to keep it from descending into `.obsidian` and `Attachments` dirs.