obs-filter-vault.py: Added attachments handling, enabled with the "--attachments" option.
When the --attachments option is used, obs-filter-vault.py will now copy (not move!) all attachments that it recognizes as being linked-to from within the output Obsidian notes files.
Moving attachments is not implemented because doing it safely would require checking _all_ documents in the source Vault and copying (rather than moving) those, so as not to create broken links.

There are *definitely* some improvements and refactoring that can be done to this.